New starry-vim !! First Version!
Using space-vim as the configuration framework.

Before and User configurations move to ~/.starry file.
Local configurations move to local/ directory.
Fork configurations move to fork/ directory.

+++ uninstall.sh
@@ -1,21 +1,25 @@
-#!/usr/bin/env sh
+#!/bin/bash
 
 app_dir="$HOME/.starry-vim"
 app_backup_dir="$HOME/.cache/.starry-vim_backup"
-
-warn() {
-    echo "$1" >&2
-}
-
-die() {
-    warn "$1"
-    exit 1
-}
+answer="n"
 
 rm $HOME/.vimrc
-rm $HOME/.vimrc.before
-rm $HOME/.vimrc.plugs
-rm -rf $HOME/.vim
 
 rm -rf $app_dir
 rm -rf $app_backup_dir
+
+printf "\rRemove the file .starry and dir .vim?\n"
+for (( i=10; i>=0; i-- )); do
+    printf "\r[y(es)/n(o), default: n]( ${i}s ): "
+    read -n 1 -t 1 answer
+    if [ "$?" -eq 0 ]; then
+        break
+    fi
+done
+if [[ "$answer" =~ ^[yY]$ ]]; then
+    rm $HOME/.starry
+    rm -rf $HOME/.vim
+fi
+
+echo "Done."
